TN-1751 followup validator code cleanup
